requirements.txt
streamlit~=1.42.0
pillow~=11.1.0
protobuf~=5.29.3
selenium~=4.29.0
webdriver-manager~=4.0.2